complete refactor of the code, create StatusError for error handling, using Bindingresult with validation
