+++ src/main/java/com/greenfox/poker/controller/RestController.java
@@ -5,6 +5,7 @@ import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -16,18 +17,12 @@ public class RestController {
   UserService userService;
 
   @RequestMapping(value = "/register", method = RequestMethod.POST)
-  public ResponseEntity<?> register(@RequestBody @Valid PokerUser userRegister) {
-    if(userService.validateUserRegistration(userRegister)){
-      return new ResponseEntity(userService.mockResponse(), HttpStatus.OK );
-    }else
-    return new ResponseEntity(userService.respondToRegisterError("mivan"), HttpStatus.BAD_REQUEST);
+  public ResponseEntity<?> register(@RequestBody @Valid PokerUser userRegister, BindingResult bindingResult) {
+    return userService.createResponseJson(bindingResult);
   }
 
   @RequestMapping(value = "/login", method = RequestMethod.POST)
-  public ResponseEntity<?> login(@RequestBody @Valid PokerUser userLogin) {
-    if(userService.validateUserLogin(userLogin)){
-      return new ResponseEntity(userService.loginUser(userLogin), HttpStatus.OK );
-    }else
-      return new ResponseEntity(userService.respondToLoginError("login elbaszva"), HttpStatus.BAD_REQUEST);
+  public ResponseEntity<?> login(@RequestBody @Valid PokerUser userLogin, BindingResult bindingResult) {
+    return userService.createResponseJson(bindingResult);
   }
 }
